requirements.txt
# Minimum Django, REST framework and inflection version
Django>=1.6,<1.9.0
djangorestframework>=3.1.0
inflection==0.3.1

# Test requirements
pytest-django==2.8.0
pytest==2.7.2
pytest-cov==1.6
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.14.0
